+++ test/communicator_scatter.cpp
@@ -14,14 +14,14 @@ auto main() -> int try {
 	mpi3::communicator world = env.world();
 
 	{
-		std::vector<double> v(world.size());
+		std::vector<double> v(static_cast<std::size_t>(world.size()));
 		iota(begin(v), end(v), 0);
 		std::vector<double> w(1);
 		world.scatter(begin(v), end(v), begin(w), 0);
 		assert( w[0] == world.rank() );
 	}
 	{
-		std::vector<double> v(world.root()?world.size():0);
+		std::vector<double> v(world.root()?static_cast<std::size_t>(world.size()):0);
 		iota(begin(v), end(v), 0);
 		std::vector<double> w(1);
 	//	auto e = 
@@ -30,7 +30,7 @@ auto main() -> int try {
 		assert( w[0] == world.rank() );
 	}
 	{
-		std::vector<double> v(world.root()?world.size():0);
+		std::vector<double> v(world.root()?static_cast<std::size_t>(world.size()):0);
 		iota(begin(v), end(v), 0);
 		double w = -1;
 	//	auto e = 
@@ -39,19 +39,19 @@ auto main() -> int try {
 		assert( w == world.rank() );
 	}
 	{
-		std::vector<double> v(world.root()?world.size():0);
+		std::vector<double> v(world.root()?static_cast<std::size_t>(world.size()):0);
 		iota(begin(v), end(v), 0);
 		double w = world.scatter(begin(v), end(v), 0);
 		assert( w == world.rank() );
 	}
 	{
-		std::vector<double> v(world.root()?world.size():0);
+		std::vector<double> v(world.root()?static_cast<std::size_t>(world.size()):0);
 		iota(begin(v), end(v), 0);
 		double w = v / world;
 		assert( w == world.rank() );
 	}
 	{
-		std::list<double> l(world.root()?world.size():0);
+		std::list<double> l(world.root()?static_cast<std::size_t>(world.size()):0);
 		iota(begin(l), end(l), 0);
 		double w = l / world;
 		assert( w == world.rank() );
